cinema-app/users-service/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
email-validator==2.1.0
